requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
psycopg2-binary>=2.9
django-cors-headers>=4.0
django-filter>=23.0
Markdown>=3.4
django-jazzmin>=3.0.0
openai>=1.0.0
requests>=2.31.0
pinyin
